Add usage instructions for CI-built container
I'm a little bit hazy on how this works, but this should work, if I'm
not entirely mistaken. Not sure how we can list all the tagged images or
if that's best practice. Maybe our runscript should also have a
parameter that specifies the image name. Let's keep it like this for
now.